+++ app/gui/src/components/ProjectLog/projectLogData.ts
