+++ app/gui/tsconfig.app.json
@@ -34,10 +34,12 @@
     "./src/components/CommandPalette.vue",
     "./src/components/LoadingScreen.vue",
     "./src/components/ProjectLog.vue",
+    "./src/components/ProjectLog/ProjectLogTable.vue",
+    "./src/components/ProjectLog/ProjectLogToolbar.vue",
+    "./src/components/ProjectLog/projectLogData.ts",
     "./src/components/ProtectedLayout.vue",
     "./src/components/RegistrationPage.vue",
     "./src/components/WithCurrentProject.vue",
-    "./src/components/projectLogData.ts",
     "./src/composables/appTitle.ts",
     "./src/composables/paywall/FeaturesConfiguration.ts",
     "./src/composables/paywall/index.ts",
